fix: added Automatic-Module-Name to MANIFEST.MF
When the Automatic-Module-Name entry is not added to the jar's
manifest file, Maven issues a warning during the compilation of a
project that uses that jar.

Maven issues following warning:
"Required filename-based automodules detected: [msal4j-1.13.0.jar].
Please don't publish this project to a public artifact repository!"

Currently, only a warning is generated. However, this might cause
problems in the future, especially when a project embraces the
Java Platform Module System (JPMS).