Fix PKCS12 certs not allowed to contain CA certs
It is a common scenario that a PKCS12 keystore contains only one private key, but multiple aliases/entries: for example we need to import the CA certs to the keystore before we could import the renewed cert.
So the idea is to only look for private key type entires in the keystore and do not throw an exception if it also contains CA or other non-private key certs, but only if it contains zero or multiple private keys as those scenarios are indeed invalid.

Note that I had trouble mocking the java KeyStore aliases() method in the unit tests with @PrepareForTest({KeyStore.class}) + EasyMock/PowerMock.createMock even tough PowerMock should support mocking final methods, but I could mock the KeyStoreSpi.